+++ irc/protocol.d
@@ -22,7 +22,7 @@ struct IrcLine
  */
 class IrcParseErrorException : Exception
 {
-	this(string msg, string file = __FILE__, uint line = __LINE__)
+	this(string msg, string file = __FILE__, size_t line = __LINE__)
 	{
 		super(msg, file, line);
 	}
